code/requirements.txt
albumentations==1.1.0
librosa==0.9.1
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
pycave==2.0.5
python_speech_features==0.6
pytorch_lightning==1.5.10
PyYAML==6.0
requests==2.27.1
samplerate==0.1.0
scikit_learn==1.0.2
scikit_plot==0.3.7
scipy==1.7.3
seaborn==0.11.2
SoundFile==0.10.3.post1
torch==1.8.1+cu101
torchvision==0.9.1+cu101
tqdm==4.63.0
